feat: add new @auth/sveltekit server side auth setup
