+++ src/routes/signout/+page.server.ts
@@ -1,0 +1,4 @@
+import { signOut } from "../../auth"
+import type { Actions } from "./$types"
+
+export const actions = { default: signOut } satisfies Actions
