+++ src/routes/signout/+page.svelte
@@ -1,0 +1,1 @@
+<!-- empty -->
